requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==4.2.27
django-extensions==4.1
django-jazzmin==3.0.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.0
python-dateutil==2.9.0.post0
reportlab==4.4.9
six==1.17.0
sqlparse==0.5.5
